+++ NEWS.md
@@ -15,6 +15,7 @@
 * `autoRecode()` for automatically recoding (numerical) variables
 
 ## bug fixes
+* bug fix in `multiChar2fac()` for cases with existing value labels (with values labeled greater than zero)
 * `composeVar()` now correctly prioritizes if both variables have missing values
 * `changeMissings()` now does no longer add arbitrary value labels for newly tagged values
 * bug fix in `checkUniqueness2()` for cases with differing numbers of cases per imputed data set